+++ example/lib/main.dart
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-import 'package:example/src/r.dart';
+import 'package:example/src/koo.dart';
 import 'package:flutter/material.dart';
 
 void main() => runApp(const MyApp());